Added breakaway options to platforms
Expanded PlatformMovement into PlatformBehavior
